Get PowerStatus and only send TurnOff if the TV is Active
Modern OLED TVs may still be reachable over the network for a while
after turning off; it's goes to "Active Standby" mode.  In this state
this function actually turns the TV back on!  So we check the state
and only send an Off signal if the TV is currently Active